requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
PyYAML==6.0.2
pydantic==2.9.2
tenacity==9.0.0
pytest==8.3.3
